requirements.txt
torch==2.1.0
torch-geometric==2.4.0
streamlit==1.29.0
Pillow==10.0.0
scikit-learn==1.3.0
numpy==1.24.3
torchvision==0.16.0
tqdm==4.65.0